+++ src/rules/require/luau_path_locator.rs
@@ -1,4 +1,7 @@
-use std::path::{Path, PathBuf};
+use std::{
+    cell::RefCell,
+    path::{Path, PathBuf},
+};
 
 use super::{path_iterator, LuauRequireMode};
 use crate::rules::require::hybrid_path_locator::SingularPathLocator;
@@ -11,7 +14,7 @@ use crate::{utils, DarkluaError, Resources};
 /// the behavior defined in the Luau RFCs for module path resolution.
 #[derive(Clone, Debug)]
 pub(crate) struct LuauPathLocator<'b, 'resources> {
-    luau_require_mode: LuauRequireMode,
+    luau_require_mode: RefCell<LuauRequireMode>,
     extra_module_relative_location: &'b Path,
     resources: &'resources Resources,
 }
@@ -23,10 +26,16 @@ impl<'b, 'c> LuauPathLocator<'b, 'c> {
         resources: &'c Resources,
     ) -> Self {
         Self {
-            luau_require_mode,
+            luau_require_mode: RefCell::new(luau_require_mode),
             extra_module_relative_location,
             resources,
         }
+    }
+
+    fn load_aliases(&self, source: &Path) -> Result<(), DarkluaError> {
+        self.luau_require_mode
+            .borrow_mut()
+            .load_aliases(source, self.resources)
     }
 }
 
@@ -35,16 +44,8 @@ impl super::PathLocator for LuauPathLocator<'_, '_> {
         &self,
         call: &crate::nodes::FunctionCall,
         _source: &Path,
-    ) -> Option<(PathBuf, SingularPathLocator<'_, '_, '_>)> {
+    ) -> Option<(PathBuf, SingularPathLocator<'_, '_>)> {
         match_path_require_call(call).map(|x| (x, SingularPathLocator::Luau(self.clone())))
-    }
-
-    fn initialize(
-        &mut self,
-        path: &Path,
-        resources: &crate::Resources,
-    ) -> Result<(), DarkluaError> {
-        self.luau_require_mode.load_aliases(path, resources)
     }
 
     fn find_require_path(
@@ -58,9 +59,12 @@ impl super::PathLocator for LuauPathLocator<'_, '_> {
             path.display(),
             source.display()
         );
+        self.load_aliases(source)?;
+
+        let require_mode = self.luau_require_mode.borrow();
 
         if is_require_relative(&path) {
-            if self.luau_require_mode.is_module_folder_name(source)? {
+            if require_mode.is_module_folder_name(source)? {
                 path = get_relative_parent_path(get_relative_parent_path(source)).join(path);
             } else {
                 path = get_relative_parent_path(source).join(path);
@@ -80,8 +84,7 @@ impl super::PathLocator for LuauPathLocator<'_, '_> {
             if source_name == "@self" {
                 path = get_relative_parent_path(source).join(components);
             } else if source_name.starts_with("@") {
-                let mut extra_module_location = self
-                    .luau_require_mode
+                let mut extra_module_location = require_mode
                     .get_source(source_name, self.extra_module_relative_location)
                     .ok_or_else(|| {
                         DarkluaError::invalid_resource_path(
@@ -95,10 +98,9 @@ impl super::PathLocator for LuauPathLocator<'_, '_> {
         }
 
         let normalized_path = utils::normalize_path_with_current_dir(&path);
-        for potential_path in path_iterator::find_require_paths(
-            &normalized_path,
-            self.luau_require_mode.module_folder_name(),
-        ) {
+        for potential_path in
+            path_iterator::find_require_paths(&normalized_path, require_mode.module_folder_name())
+        {
             if self.resources.is_file(&potential_path)? {
                 return Ok(utils::normalize_path_with_current_dir(potential_path));
             }
@@ -109,7 +111,7 @@ impl super::PathLocator for LuauPathLocator<'_, '_> {
                 "tried `{}`",
                 path_iterator::find_require_paths(
                     &normalized_path,
-                    self.luau_require_mode.module_folder_name(),
+                    require_mode.module_folder_name(),
                 )
                 .map(|potential_path| potential_path.display().to_string())
                 .collect::<Vec<_>>()