+++ src/rules/require/luau_require_mode.rs
@@ -4,8 +4,8 @@ use crate::frontend::DarkluaResult;
 use crate::nodes::{Arguments, FunctionCall, StringExpression};
 use crate::rules::require::{match_path_require_call, path_utils, LuauPathLocator, PathLocator};
 use crate::rules::{Context, RequireModeLike, SingularRequireMode};
-use crate::utils;
 use crate::DarkluaError;
+use crate::{utils, Resources};
 
 use std::collections::HashMap;
 use std::ffi::OsStr;
@@ -38,34 +38,9 @@ impl Default for LuauRequireMode {
     }
 }
 
-impl LuauRequireMode {
-    /// Attempt to load any .luaurc aliases from a source luau path.
-    pub fn load_aliases(
-        &mut self,
-        luau_file: &Path,
-        resources: &crate::Resources,
-    ) -> Result<(), DarkluaError> {
-        if let Some(config) = utils::find_luau_configuration(luau_file, resources)? {
-            self.luau_rc_aliases
-                .get_or_insert_with(HashMap::new)
-                .extend(config.aliases);
-        }
-
-        Ok(())
-    }
-}
-
 impl RequireModeLike for LuauRequireMode {
     fn initialize(&mut self, context: &Context) -> Result<(), DarkluaError> {
-        if !self.use_luau_configuration {
-            self.luau_rc_aliases.take();
-            return Ok(());
-        }
-
-        // Load aliases from .luaurc configuration
-        self.load_aliases(context.current_path(), context.resources)?;
-
-        Ok(())
+        self.load_aliases(context.current_path(), context.resources())
     }
 
     fn is_module_folder_name(&self, path: &Path) -> DarkluaResult<bool> {
@@ -228,6 +203,25 @@ impl RequireModeLike for LuauRequireMode {
 }
 
 impl LuauRequireMode {
+    pub(crate) fn load_aliases(
+        &mut self,
+        luau_file: &Path,
+        resources: &Resources,
+    ) -> Result<(), DarkluaError> {
+        if !self.use_luau_configuration {
+            return Ok(());
+        }
+
+        // Load aliases from .luaurc configuration
+        if let Some(config) = utils::find_luau_configuration(luau_file, resources)? {
+            self.luau_rc_aliases.replace(config.aliases);
+        } else {
+            self.luau_rc_aliases.take();
+        }
+
+        Ok(())
+    }
+
     /// Set if the require mode should use `.luaurc` configuration to resolve aliases.
     pub fn with_configuration(mut self, use_luau_configuration: bool) -> Self {
         self.use_luau_configuration = use_luau_configuration;