+++ tests/test_cases/small_bundle_multi_luaurc/main.lua
@@ -1,0 +1,7 @@
+local tryValueDefaultModule = require("./value")
+
+local tryValue1Module = require("./try-1/value")
+
+local tryValue2Module = require("./try-2/value")
+
+print(tryValueDefaultModule, tryValue1Module, tryValue2Module)