+++ tests/test_cases/small_bundle_multi_luaurc/try-2/value.lua
@@ -1,0 +1,3 @@
+local tryValue2 = require('@value')
+
+return tryValue2