+++ tests/test_cases/small_bundle_multi_luaurc/value.lua
@@ -1,0 +1,3 @@
+local tryValueDefault = require('@value')
+
+return tryValueDefault